rag/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
qdrant-client>=1.6.0
openai>=1.0.0
grpcio>=1.59.0
grpcio-tools>=1.59.0
protobuf>=4.25.0
tenacity>=8.2.0
python-dotenv>=1.0.0

